Moved method in IndexUtil
